@@ -8,51 +8,48 @@
 
 # -- Path setup --------------------------------------------------------------
 
-# If extensions (or modules to document with autodoc) are in another directory,
-# add these directories to sys.path here. If the directory is relative to the
-# documentation root, use os.path.abspath to make it absolute, like shown here.
-#
+# To link modules with autodoc
 import os
 import sys
 sys.path.insert(0, os.path.abspath('../../'))
 
 # -- Project information -----------------------------------------------------
-
 project = 'MAICoS'
 copyright = 'All source code is available under the GNU General Public License v3.0'
 author = 'See the file AUTHORS for the full list of names.'
 
 # The short X.Y version
-version = '0.4'
+# NOTE: keep in sync with VERSION in setup.py
+# NOTE: keep in sync with __version__ in maicos.__init__.py
+version = '0.5'
 # The full version, including alpha/beta/rc tags
-release = '0.4'
+release = version
 
 
 # -- General configuration ---------------------------------------------------
 
-# If your documentation needs a minimal Sphinx version, state it here.
-#
-# needs_sphinx = '1.0'
-
-# Add any Sphinx extension module names here, as strings. They can be
-# extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
-# ones.
+# Add any Sphinx extension module names here, as strings.
 extensions = [
-    'sphinx.ext.autodoc',
-    'sphinx.ext.viewcode',
-#    'sphinxarg.ext',
-#    'sphinx.ext.napoleon',
-#    'sphinx.ext.autosectionlabel',  
+    'sphinx.ext.autodoc', # import the modules you are documenting
+    'sphinx.ext.viewcode', # tries to find the source files where the objects are contained
+    'sphinx.ext.intersphinx', # generate links to the documentation of objects in external projects
+    # 'sphinx.ext.napoleon', # Support for NumPy and Google style docstrings
+    'nbsphinx', #  provides a source parser for *.ipynb files
 ]
+
+# Execute the notebooks
+nbsphinx_execute = 'always'
+nbsphinx_allow_errors = True
+exclude_patterns = ['_build', '**.ipynb_checkpoints']
 
 # Add any paths that contain templates here, relative to this directory.
 templates_path = ['_templates']
 
 # The suffix(es) of source filenames.
 # You can specify multiple suffix as a list of string:
-#
-# source_suffix = ['.rst', '.md']
-source_suffix = '.rst'
+# note: do not add .ipynb when nbspinx is enabled, 
+# otherwise you get the "missing title" error
+source_suffix = ['.rst']
 
 # The master toctree document.
 master_doc = 'index'
@@ -70,29 +67,26 @@ language = None
 exclude_patterns = []
 
 # The name of the Pygments (syntax highlighting) style to use.
-#pygments_style = None
-pygments_style = 'sphinx'
+pygments_style = 'tango'
 
 # -- Options for HTML output -------------------------------------------------
 
 # The theme to use for HTML and HTML Help pages.  See the documentation for
 # a list of builtin themes.
-#
-html_theme = 'press'
+html_theme = 'furo'
 
 # Theme options are theme-specific and customize the look and feel of a theme
 # further.  For a list of options available for each theme, see the
 # documentation.
 #
 # html_theme_options = {}
+html_logo = "images/logo_MAICOS.png"
 
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
 #html_static_path = ['_static']
 html_static_path = []
-
-html_logo = "images/logo_MAICOS.png"
 
 # Custom sidebar templates, must be a dictionary that maps document names
 # to template names.
@@ -103,6 +97,12 @@ html_logo = "images/logo_MAICOS.png"
 # 'searchbox.html']``.
 #
 # html_sidebars = {}
+
+#Controls whether the user can navigate the documentation 
+#using the keyboard’s left and right arrows. The default is False.
+html_theme_options = {
+    "navigation_with_keys": True,
+}
 
 
 # -- Options for HTMLHelp output ---------------------------------------------
