@@ -1,180 +1,214 @@
 #!/usr/bin/env python3
 # -*- Mode: python; tab-width: 4; indent-tabs-mode:nil; coding:utf-8 -*-
 #
-# Copyright (c) 2021 Authors and contributors
+# Copyright (c) 2022 Authors and contributors
 # (see the file AUTHORS for the full list of names)
 #
 # Released under the GNU Public Licence, v2 or any higher version
 # SPDX-License-Identifier: GPL-2.0-or-later
 
 import logging
-import warnings
 from ..utils import sort_atomgroup
 
 import numpy as np
-from MDAnalysis.analysis import base
-from MDAnalysis.lib.log import ProgressBar
+import MDAnalysis.analysis.base
+
+from ..decorators import set_verbose_doc, set_planar_class_doc
 
 logger = logging.getLogger(__name__)
 
+@set_verbose_doc
+class AnalysisBase(MDAnalysis.analysis.base.AnalysisBase):
+    """Base class derived from MDAnalysis for defining multi-frame analysis.
 
-class _AnalysisBase(base.AnalysisBase):
-    """Extends the MDAnalysis base class for defining multi frame analysis."""
+    The class is designed as a template for creating multi-frame analyses.
+    This class will automatically take care of setting up the trajectory
+    reader for iterating, and it offers to show a progress meter.
+    Computed results are stored inside the :attr:`results` attribute.
 
-    def __init__(self, trajectory, verbose=False, save=False, **kwargs):
-        """
-        Parameters
-        ----------
-        trajectory : mda.Reader
-            A trajectory Reader
-        verbose : bool, optional
-           Turn on more logging and debugging, default ``False``
-        save : bool, optional
-           Save results to a file, default ``False``
-        """
-        super().__init__(self, **kwargs)
+    To define a new analysis, `AnalysisBase` needs to be subclassed
+    and :meth:`_single_frame` must be defined. It is also possible to define
+    :meth:`_prepare` and :meth:`_conclude` for pre- and post-processing.
+    All results should be stored as attributes of the :class:`Results`
+    container.
 
-        self._trajectory = trajectory
-        self._verbose = verbose
-        self._save = save
-        self.results = {}
+    Parameters
+    ----------
+    atomgroups : Atomgroup or list[Atomgroup]
+        Atomgroups taken for the Analysis
+    multi_group : bool
+        Analysis is able to work with list of atomgroups
+    ${VERBOSE_PARAMETER}
 
-    def _setup_frames(self, trajectory, begin=0, end=None, dt=0):
-        """
-        Pass a Reader object and define the desired iteration pattern
-        through the trajectory
-
-        Parameters
-        ----------
-        trajectory : mda.Reader
-            A trajectory Reader
-        begin : float, optional
-            start time (ps) for evaluation
-        end : float, optional
-            end time (ps) for evaluation
-        dt : float, optional
-            time step (ps) to read analysis frame. If `0` take all frames
-        """
-        self._trajectory = trajectory
-
-        self.begin = begin
-        self.end = end
-        self.dt = dt
-
-        with warnings.catch_warnings():
-            warnings.simplefilter('always')
-            if begin > trajectory.totaltime:
-                raise ValueError("Start ({:.2f} ps) is larer than total time "
-                                 "({:.2f} ps)."
-                                 "".format(begin, trajectory.totaltime))
-
-            if end is not None and end < trajectory.dt:
-                raise ValueError("End ({:.2f} ps) is smaller than a single "
-                                 "trajectory timestep ({:.2f} ps)."
-                                 "".format(end, trajectory.dt))
-
-            if begin > 0:
-                startframe = int(begin // trajectory.dt)
-            else:
-                startframe = 0
-            if end is not None:
-                stopframe = int(end // trajectory.dt)
-            else:
-                stopframe = trajectory.n_frames
-            if self.dt > 0:
-                step = int(dt // trajectory.dt)
-            else:
-                step = 1
-
-        startframe, stopframe, step = trajectory.check_slice_indices(
-            startframe, stopframe, step)
-        self.startframe = startframe
-        self.stopframe = stopframe
-        self.step = step
-        self.n_frames = len(range(startframe, stopframe, step))
-        self.frames = np.zeros(self.n_frames, dtype=int)
-        self.times = np.zeros(self.n_frames)
-
-    def _configure_parser(self, parser):
-        """Adds parser options using an argparser object"""
-        parser.description = self.__doc__
-
-    def _calculate_results(self):
-        """Calculate the results"""
-        pass
-
-    def _save_results(self):
-        """Saves the results you've gatherd to a file."""
-        pass
-
-    def run(self, begin=0, end=None, dt=0, verbose=None):
-        """Perform the calculation
-
-        Parameters
-        ----------
-        begin : float, optional
-            start time (ps) for evaluation
-        end : float, optional
-            end time (ps) for evaluation
-        dt : float, optional
-            time step (ps) to read analysis frame
-        verbose : bool, optional
-            Turn on verbosity
-        """
-        logger.info("Choosing frames to analyze")
-        # if verbose unchanged, use class default
-        verbose = getattr(self, '_verbose',
-                          False) if verbose is None else verbose
-
-        self._setup_frames(self._trajectory, begin, end, dt)
-        logger.info("Starting preparation")
-        self._prepare()
-        for i, ts in enumerate(ProgressBar(
-                self._trajectory[self.startframe:self.stopframe:self.step],
-                verbose=verbose)):
-            self._frame_index = i
-            self.frames[i] = ts.frame
-            self.times[i] = ts.time
-            self._ts = ts
-            # logger.info("--> Doing frame {} of {}".format(i+1, self.n_frames))
-            self._single_frame()
-        logger.info("Finishing up")
-        self._calculate_results()
-        self._conclude()
-        if self._save:
-            self._save_results()
-        return self
-
-
-class SingleGroupAnalysisBase(_AnalysisBase):
-    """The base class for analysing a single AtomGroup only."""
-
-    _allow_multiple_atomgroups = False
-
-    def __init__(self, atomgroup, **kwargs):
-        super().__init__(atomgroup.universe.trajectory, **kwargs)
-        self.atomgroup = sort_atomgroup(atomgroup)
-        self._universe = atomgroup.universe
-
-
-class MultiGroupAnalysisBase(_AnalysisBase):
-    """The base class for analysing a single or multiple AtomGroups."""
-
-    _allow_multiple_atomgroups = True
-
-    def __init__(self, atomgroups, **kwargs):
-        if type(atomgroups) not in [list, tuple, np.ndarray]:
-            # Sort the atomgroups,
-            # such that molecules are listed one after the other
-            atomgroups = [sort_atomgroup(atomgroups)]
-        else:
-            atomgroups = list(map(sort_atomgroup, atomgroups))
+    Attributes
+    ----------
+    atomgroup : mda.Atomgroup
+        Atomgroup taken for the Analysis (available if `multi_group = False`)
+    atomgroups : list[mda.Atomgroup]
+        Atomgroups taken for the Analysis (available if `multi_group = True`)
+    n_atomgroups : int
+        Number of atomngroups (available if `multi_group = True`)
+    _universe : mda.Universe
+        The Universe the atomgroups belong to
+    _trajectory : mda.trajectory
+        The trajetcory the atomgroups belong to
+    times : numpy.ndarray
+        array of Timestep times. Only exists after calling
+        :meth:`AnalysisBase.run`
+    frames : numpy.ndarray
+        array of Timestep frame indices. Only exists after calling
+        :meth:`AnalysisBase.run`
+    results : :class:`Results`
+        results of calculation are stored after call
+        to :meth:`AnalysisBase.run`
+    """
+    def __init__(self,
+                 atomgroups,
+                 multi_group=False,
+                 verbose=False,
+                 **kwargs):
+        if multi_group:
+            if type(atomgroups) not in (list, tuple):
+                atomgroups = [atomgroups]
             # Check that all atomgroups are from same universe
             if len(set([ag.universe for ag in atomgroups])) != 1:
-                raise ValueError(
-                    "Given Atomgroups are not from the same Universe.")
-        super().__init__(atomgroups[0].universe.trajectory, **kwargs)
+                raise ValueError("Atomgroups belong to different Universes")
 
-        self.atomgroups = atomgroups
-        self.n_atomgroups = len(self.atomgroups)
-        self._universe = atomgroups[0].universe
+            # Sort the atomgroups,
+            # such that molecules are listed one after the other
+            self.atomgroups = list(map(sort_atomgroup, atomgroups))
+            self.n_atomgroups = len(self.atomgroups)
+            self._universe = atomgroups[0].universe
+            self._allow_multiple_atomgroups = True
+        else:
+            self.atomgroup = sort_atomgroup(atomgroups)
+            self._universe = atomgroups.universe
+            self._allow_multiple_atomgroups = False
+
+        self._trajectory = self._universe.trajectory
+
+        super(AnalysisBase, self).__init__(trajectory=self._trajectory,
+                                           verbose=verbose,
+                                           **kwargs)
+
+@set_planar_class_doc
+class PlanarBase(AnalysisBase):
+    """Class to provide options and attributes for analysis in planar system.
+
+    Provied the results attribute `z`.
+
+    Parameters
+    ----------
+    trajectory : MDAnalysis.coordinates.base.ReaderBase
+        A trajectory Reader
+    ${PLANAR_CLASS_PARAMETERS}
+    kwargs : dict
+        Parameters parsed to `AnalysisBase`.
+
+    Attributes
+    ----------
+    ${PLANAR_CLASS_ATTRIBUTES}
+    zmin : float
+        Minimal position for analysis (Å)
+    zmax : float
+        Maximal position for analysis (Å)
+    n_bins : int
+        Number of bins for analysis
+
+    """
+
+    def __init__(self, atomgroups, dim, binwidth, comgroup, center, **kwargs):
+        super(PlanarBase, self).__init__(atomgroups, **kwargs)
+        self.dim = dim
+        self.binwidth = binwidth
+        self.center = center
+        self.comgroup = comgroup
+
+    def _prepare(self):
+        """Preparations for the planar analysis."""
+        if self.dim not in [0, 1, 2]:
+            raise ValueError("Dimension can only be x=0, y=1 or z=2.")
+
+        # Workaround since currently not alle module have option
+        # with zmax and zmin
+        if not hasattr(self, '_zmax'):
+            self._zmax = None
+
+        if self._zmax is None:
+            self.Lz = 0
+            self.zmax = self._universe.dimensions[self.dim]
+        else:
+            self.zmax = 10 * self._zmax
+
+        if not hasattr(self, 'zmin'):
+            self.zmin = 0
+        self.zmin *= 10
+        self.binwidth *= 10
+
+        self.n_bins = int(np.ceil((self.zmax - self.zmin) / self.binwidth))
+
+        logger.info(f"Using {self.n_bins} bins")
+
+        if self.comgroup is not None and self.comgroup.n_atoms == 0:
+            raise ValueError(f"`Comgroup` does not contain any atoms.")
+        if self.comgroup is not None:
+            self.center = True  # always center when COM
+
+    def get_bins(self, positions, dim=None):
+        """"Calculates bins based on given positions. dim denotes
+        the dimension for calculating bins. If `None` the default
+        dim is taken.
+
+        Attributes
+        ----------
+        positions : numpy.ndarray
+            3 dimensional positions
+        dim : int
+            dimension for binning (x=0, y=1, z=2)
+
+        Returns
+        -------
+        numpy.ndarray
+            binned psoitions
+        """
+        dim = self.dim if dim is None else dim
+        dz = self._ts.dimensions[dim] / self.n_bins
+        bins = np.rint(positions[:, dim] / dz)
+        bins %= self.n_bins
+        return bins.astype(int)
+
+    def _single_frame(self):
+        """Single frame for the planar analysis."""
+        if self._zmax is None:
+            self.zmax = self._ts.dimensions[self.dim]
+            self.Lz += self.zmax
+
+        if self.comgroup is not None:
+            comshift = self.comgroup.center_of_mass(pbc=True)
+
+            if hasattr(self, 'atomgroup'):
+                groups = [self.atomgroup]
+            else:
+                groups = self.atomgroups
+            for group in groups:
+                group.atoms.positions += comshift
+
+    def _conclude(self):
+        """Results calculations for the planar analysis."""
+        self._index = self._frame_index + 1
+
+        if self._zmax is None:
+            zmax = self.Lz / self._index
+        else:
+            zmax = self.zmax
+
+        dz = (zmax - self.zmin) / self.n_bins
+
+        self.results.z = np.linspace(
+            self.zmin+dz/2, zmax-dz/2, self.n_bins,
+            endpoint=False)
+
+        if self.center:
+            self.results.z -= self.zmin + (zmax - self.zmin) / 2
+
+        self.results.z /= 10
