@@ -19,5 +19,13 @@ AIRWATER_GRO = resource_filename(__name__, "../data/airwater/confout.gro")
 AIRWATER_TRR = resource_filename(__name__, "../data/airwater/traj.trr")
 AIRWATER_TPR = resource_filename(__name__, "../data/airwater/topol.tpr")
 
-SALT_WATER_GRO = resource_filename(__name__, "../data/salt_water/salt_water.gro")
-SALT_WATER_TPR = resource_filename(__name__, "../data/salt_water/salt_water.tpr")
+SALT_WATER_GRO = resource_filename(
+    __name__, "../data/salt_water/salt_water.gro")
+SALT_WATER_TPR = resource_filename(
+    __name__, "../data/salt_water/salt_water.tpr")
+
+# NVE bulk water
+NVE_WATER_TPR = resource_filename(__name__, "../data/kineticenergy/nve.tpr")
+NVE_WATER_TRR = resource_filename(__name__, "../data/kineticenergy/nve.trr")
+
+LAMMPS10WATER = resource_filename(__name__, "../data/lammps10water.data")
